ml/requirements.txt
# ML Training Dependencies
# Recommended: Use Python 3.12 venv (Python 3.14 has compatibility issues with TF/albumentations)
#   python -m uv venv .venv312 --python 3.12
#   python -m uv pip install --python .venv312/Scripts/python.exe -r requirements.txt

torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=9.0.0
albumentations>=1.3.0
tqdm>=4.65.0
pyyaml>=6.0
scipy>=1.10.0

# ONNX export
onnx>=1.14.0
onnxscript>=0.1.0
onnxruntime>=1.15.0

# TFLite conversion (requires Python <=3.12)
onnx2tf>=1.20.0
tensorflow>=2.15.0
